+++ udemy/angular_spring_boot/helpdesk/src/main/java/com/aug/helpdesk/repositories/ClienteRepository.java
@@ -1,11 +1,9 @@
 package com.aug.helpdesk.repositories;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aug.helpdesk.domain.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, BigInteger> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
 
 }
